requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
gspread>=5.12.0
oauth2client>=4.1.3
numpy>=1.24.0
PyYAML>=6.0.0
